+++ lib/browser/browser.ts
@@ -111,7 +111,6 @@ class OktaAuthBrowser extends OktaAuthBase implements OktaAuth, SignoutAPI {
       redirectUri: args.redirectUri,
       postLogoutRedirectUri: args.postLogoutRedirectUri,
       responseMode: args.responseMode,
-      transformErrorXHR: args.transformErrorXHR,
       cookies: getCookieSettings(this, args)
     });