+++ CHANGELOG.md
@@ -6,6 +6,8 @@
 
 - [#489](https://github.com/okta/okta-auth-js/pull/489) Fixes sameSite cookie setting when running on HTTP connection
 
+- [#468](https://github.com/okta/okta-auth-js/pull/468) Fixes issue where HTTP headers with an undefined value were being sent with the value "undefined". These headers are now removed before the request is sent.
+
 ## 4.0.1
 
 ### Bug Fixes